requirements.txt
asgiref==3.3.0
autopep8==1.5.4
beautifulsoup4==4.9.3
Django==3.0.2
django-bootstrap4==2.3.1
django-crispy-forms==1.9.2
django-filter==2.4.0
django-tinymce==3.1.0
djangorestframework==3.12.1
Markdown==3.3.3
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
pytz==2020.4
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
